# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API backend",
        "description": "Documentação API backend - CRUD",
        "contact": {
            "email": "assuncaofilipe97@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local"
        }
    ],
    "paths": {
        "/api/cake": {
            "post": {
                "tags": [
                    "cake"
                ],
                "summary": "Cadastrar um bolo",
                "description": "Retorna um bolo cadastrado",
                "operationId": "84f14d8ac8e3113f66e5f6e2a80838d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "$ref": "#/components/schemas/cakeRequestBody"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "register-profile": {
                                    "summary": "Cadastrar um bolo",
                                    "value": {
                                        "name": "Bolo de chocolate",
                                        "weight": "800,00",
                                        "price": "150,00",
                                        "available_quantity": "5"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cadastro de de um bolo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/cake"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "id": "2",
                                        "name": "bolo de chocolate",
                                        "weight": "150",
                                        "price": "5.00",
                                        "available_quantity": "1",
                                        "uuid": "0019b5c5-8390-4fcf-9db7-55d07a5f3860",
                                        "updated_at": "2022-03-28T22:19:57.000000Z",
                                        "created_at": "2022-03-28T22:19:57.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválido",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "name": [
                                            "O campo name é obrigatório.",
                                            "O campo name não pode ser superior a 100 caracteres."
                                        ],
                                        "weight": [
                                            "O campo weight é obrigatório.",
                                            "O campo weight não pode ser superior a 99999.",
                                            "O campo weight deve ser pelo menos 0."
                                        ],
                                        "price": [
                                            "O campo price é obrigatório.",
                                            "O campo price não pode ser superior a 99999.",
                                            "O campo price deve ser pelo menos 0."
                                        ],
                                        "available_quantity": [
                                            "O campo available quantity é obrigatório.",
                                            "O campo available quantity deve ser um número inteiro."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/cake/{id}": {
            "get": {
                "tags": [
                    "cake"
                ],
                "summary": "Listar as informações de um bolo",
                "description": "Retorna as informações de um bolo",
                "operationId": "ac5797633d2c9ea4ab1671829b584a24",
                "responses": {
                    "200": {
                        "description": "Informações de um bolo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/cake"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "id": "1",
                                        "name": "bolo de chocolate com avelã",
                                        "weight": "801",
                                        "price": "5.00",
                                        "available_quantity": "1",
                                        "uuid": "0019b5c5-8390-4fcf-9db7-55d07a5f3860",
                                        "updated_at": "2022-03-28T22:19:57.000000Z",
                                        "created_at": "2022-03-28T22:19:57.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bolo inexistente",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "message": "Bolo inexistente"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "cake"
                ],
                "summary": "Atualizar informações de um bolo",
                "description": "Retorna um bolo atualizado",
                "operationId": "3b056f3d26ef33aa7c9f19b0ed60e4a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "$ref": "#/components/schemas/cakeRequestBody"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "register-profile": {
                                    "summary": "Atualiar um bolo",
                                    "value": {
                                        "name": "Bolo de chocolate com avelã",
                                        "weight": "801,00",
                                        "price": "150,00",
                                        "available_quantity": "5"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualização de um bolo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/cake"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "id": "2",
                                        "name": "bolo de chocolate com avelã",
                                        "weight": "801",
                                        "price": "5.00",
                                        "available_quantity": "1",
                                        "uuid": "0019b5c5-8390-4fcf-9db7-55d07a5f3860",
                                        "updated_at": "2022-03-28T22:19:57.000000Z",
                                        "created_at": "2022-03-28T22:19:57.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "name": [
                                            "O campo name não pode ser superior a 100 caracteres."
                                        ],
                                        "weight": [
                                            "O campo weight não pode ser superior a 99999.",
                                            "O campo weight deve ser pelo menos 0."
                                        ],
                                        "price": [
                                            "O campo price não pode ser superior a 99999.",
                                            "O campo price deve ser pelo menos 0."
                                        ],
                                        "available_quantity": [
                                            "O campo available quantity deve ser um número inteiro."
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bolo inexistente",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "message": "Bolo inexistente"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cake"
                ],
                "summary": "Deletar um bolo",
                "description": "Retorna um bolo deletado",
                "operationId": "75c1ca0bfc3ddb8bdac1148f1340d07a",
                "responses": {
                    "200": {
                        "description": "Retorna as informações de um bolo deletado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/cake"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "id": "2",
                                        "name": "bolo de chocolate com avelã",
                                        "weight": "801",
                                        "price": "5.00",
                                        "available_quantity": "1",
                                        "uuid": "0019b5c5-8390-4fcf-9db7-55d07a5f3860",
                                        "deleted_at": "2022-03-28T22:20:57.000000Z",
                                        "updated_at": "2022-03-28T22:19:57.000000Z",
                                        "created_at": "2022-03-28T22:19:57.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bolo inexistente",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": "false",
                                    "data": {
                                        "message": "Bolo inexistente"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "cake": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "weight": {
                        "type": "number",
                        "format": "decimal(8,2)"
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal(8,2)"
                    },
                    "avalaible_quantity": {
                        "type": "integer"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "cakeRequestBody": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "weight": {
                        "type": "number",
                        "format": "decimal(8,2)"
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal(8,2)"
                    },
                    "avalaible_quantity": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}